Refactor duplicated code chunks
Remove unused generic utility and unnecessary Bot screens
