--- Task4_Battleships/GameElements/Utilities/GenericUtility.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Task4_Battleships.GameElements.Utilities
-{
-    internal class GenericUtility
-    {
-        
-
-        public GenericUtility() 
-        { 
-                    
-        }
-    }
-}
